requirements.txt
#PYTHON - 3.11
aniso8601==10.0.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
et_xmlfile==2.0.0
flasgger==0.9.7.1
Flask==3.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
marshmallow==3.26.1
mistune==3.1.3
numpy==2.2.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pyngrok==7.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
six==1.17.0
SQLAlchemy==2.0.40
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
Werkzeug==3.1.3
pyngrok==5.1.0